sized numbers

+++ test/TyperSpec.hs
@@ -14,12 +14,12 @@ spec = do
   describe "Functions" $ do
     it "should be passed the correct number of arguments" $
        let progFail = "\
-                   \int add(int x, int y) {return 0;}\n\
-                   \int main(){return add(1);}\n"
+                   \i64 add(i64 x, i64 y) {return 0;}\n\
+                   \i64 main(){return add(1);}\n"
 
            progSucc = "\
-                   \int add(int x, int y) {return 0;}\n\
-                   \int main(){return add(1, 2);}\n"
+                   \i64 add(i64 x, i64 y) {return 0;}\n\
+                   \i64 main(){return add(1, 2);}\n"
        in do
          compileString progFail `shouldSatisfy` isError
          compileString progSucc `shouldSatisfy` isSucc
